Aggiunti controlli DTO a Search Tavoli + Modifica dataType da Double a Int per la Puntata Minima

+++ src/main/java/it/pokeronline/dto/UserDTO.java
@@ -119,7 +119,7 @@ public class UserDTO {
 		result.setUsername(userDTO.getUsername());
 		result.setPassword(userDTO.getPassword());
 		result.setExpAccumulata(Long.parseLong(userDTO.getExpAccumulata()));
-		result.setCreditoAccumulato(Double.parseDouble(userDTO.getCreditoAccumulato()));
+		result.setCreditoAccumulato(Integer.parseInt(userDTO.getCreditoAccumulato()));
 		return result;
 	}
